feat: Add support for dry run

+++ main.py
@@ -8,6 +8,9 @@ import logging
 import yaml
 
 logging.basicConfig(level=logging.INFO)
+logging.getLogger("urllib3").setLevel(logging.WARNING)
+
+DRY_RUN = os.getenv("DRY_RUN", "true").lower() == "true"
 
 
 class Project(BaseModel):
@@ -19,6 +22,7 @@ class Auth:
     def __init__(self, GITHUB_TOKEN):
         self.github = Github(GITHUB_TOKEN)
         self.org = self.github.get_organization("SDC-MUJ")
+        self.logger = logging.getLogger("auth")
 
     def update_project(self, project: Project, internal_team: list[str]) -> None:
         for repo_meta in project.repos:
@@ -26,39 +30,50 @@ class Auth:
                 repo = self.org.get_repo(repo_meta)
             except UnknownObjectException:
                 # Create repo if not exist
-                logging.warning(f"Repo {repo_meta} does not exist. Creating ...")
+                self.logger.warning(f"Repo {repo_meta} does not exist. Creating ...")
+                # DO NOT create repo in dry run
+                if DRY_RUN:
+                    continue
                 repo = self.org.create_repo(name=repo_meta, private=True)
 
-            collaborators: set[str] = {
-                user.login.lower() for user in repo.get_collaborators()
-            }
-            logging.debug(f"Current collaborators for {repo.name}: {collaborators}")
+            collaborators: set[str] = {user.login.lower() for user in repo.get_collaborators()}
+            self.logger.debug(f"Current collaborators for {repo.name}: {collaborators}")
 
             # Handle new additions
             additions = set(project.maintainers) - (collaborators | set(internal_team))
             for user in additions:
-                logging.info("Adding user %s to %s", user, repo.name)
+                self.logger.info("Adding %s to %s", user, repo.name)
+                # DO NOT add collaborators in dry run
+                if DRY_RUN:
+                    continue
                 repo.add_to_collaborators(user, permission="maintain")
 
             # Warn for removals of old maintainers
             removals = collaborators - (set(project.maintainers) | set(internal_team))
             for user in removals:
-                logging.warning("Removing user %s from %s", user, repo.name)
+                self.logger.warning("Removing user %s from %s", user, repo.name)
+                # DO NOT remove collaborators in dry run
+                if DRY_RUN:
+                    continue
                 repo.remove_from_collaborators(user)
 
     def update_org(self, maintainers: set[str]) -> None:
         org_members = self.org.get_members()
         org_members_list = {member.login.lower() for member in org_members}
-        logging.debug(f"Current organization members: {org_members_list}")
+        self.logger.debug(f"Current organization members: {org_members_list}")
 
         # Handle new additions
         additions = maintainers - org_members_list
+        self.logger.debug(f"Organization additions: {additions}")
 
         for member in additions:
             user = self.github.get_user(member)
             if isinstance(user, AuthenticatedUser):
                 continue
-            logging.info("Adding %s to organization", member)
+            self.logger.info("Adding %s to organization", member)
+            # DO NOT invite users in dry run
+            if DRY_RUN:
+                continue
             self.org.invite_user(user, role="direct_member")
 
 
@@ -87,14 +102,17 @@ def parse_and_flatten(projects_yaml: str) -> Tuple[list[Project], list[str], set
 
 
 def main():
+    if DRY_RUN:
+        logging.info("DRY_RUN is activated. No write operations will be done on GitHub.")
+
     GITHUB_TOKEN = os.environ["X_GITHUB_TOKEN"]
     auth = Auth(GITHUB_TOKEN)
 
-    projects, internal_team, maintainers = parse_and_flatten("projects.yaml")
+    projects, internal_team, _ = parse_and_flatten("projects.yaml")
     for project in projects:
         auth.update_project(project, internal_team)
 
-    auth.update_org(maintainers)
+    # auth.update_org(maintainers)
 
 
 if __name__ == "__main__":
